主要定義をファイル訳
